STOR-6 move Configurations and Properties to its own Directories
